+++ src/main/java/jm/audio/io/SampleIn.java
@@ -328,7 +328,7 @@ public class SampleIn extends AudioObject implements jm.JMC{
         int ret = 0;
         int length = b.length;
         for(int i=0;i<b.length;i++,length--) {
-            ret |= ((int)(b[i] & 0xFF) << ((((bigEndian) ? length : (i+1)) * 8) - 8));
+            ret |= ((b[i] & 0xFF) << ((((bigEndian) ? length : (i+1)) * 8) - 8));
         }
         switch(sampleSize) {
         case 1:
@@ -337,7 +337,7 @@ public class SampleIn extends AudioObject implements jm.JMC{
                 ret &= 0x7F;
                 ret = ~ret + 1;
             }
-            sample = (float)((float)ret/(float)Byte.MAX_VALUE);
+            sample = (float)ret/(float)Byte.MAX_VALUE;
             break;
         case 2:
             if(ret > 0x7FFF) {
@@ -345,7 +345,7 @@ public class SampleIn extends AudioObject implements jm.JMC{
                 ret &= 0x7FFF;
                 ret = ~ret + 1;
             }
-            sample = (float)((float)ret/(float)Short.MAX_VALUE);
+            sample = (float)ret/(float)Short.MAX_VALUE;
             break;
         case 3:
             if(ret > 0x7FFFFF) {
@@ -353,7 +353,7 @@ public class SampleIn extends AudioObject implements jm.JMC{
                 ret &= 0x7FFFFF;
                 ret = ~ret + 1;
             }
-            sample = (float)((float)ret/8388608f);
+            sample = (float)ret/8388608f;
             break;
         case 4:
             sample = (float)((double)ret/(double)Integer.MAX_VALUE);
