+++ src/main/java/jm/music/tools/PhraseAnalysis.java
@@ -259,7 +259,7 @@ public final class PhraseAnalysis {
      * extremely large lowering intervals with rests.
      */
     public static final int INTERVAL_WITH_REST =
-            ((int)Note.MAX_PITCH - (int)Note.MIN_PITCH) * 2 + 1;
+            (Note.MAX_PITCH - Note.MIN_PITCH) * 2 + 1;
 
     /**
      * Integer array containing the pitches that represent 'good' intervals.  A
@@ -766,7 +766,7 @@ public final class PhraseAnalysis {
     public static double climaxStrength (final Note[] noteArray)
                                  throws NoteListException {
         int count = 0;
-        int highestPitch = (int)Note.MIN_PITCH;
+        int highestPitch = Note.MIN_PITCH;
         int currentPitch;
 
         for (int i = 0; i < noteArray.length; i++) {
@@ -1633,7 +1633,7 @@ public final class PhraseAnalysis {
                                          int[] scale)
                                  throws NoteListException {
         if (noteCount (noteArray) > 0) {
-            int highestPitch = (int)Note.MIN_PITCH;
+            int highestPitch = Note.MIN_PITCH;
             int currentPitch;
     
             for (int i = 0; i < noteArray.length; i++) {
@@ -2046,13 +2046,13 @@ public final class PhraseAnalysis {
      */
     public static int pitchRange (final Note[] noteArray)
                           throws NoteListException {
-        int highestPitch = (int)Note.MIN_PITCH;
-        int lowestPitch = (int)Note.MAX_PITCH;
+        int highestPitch = Note.MIN_PITCH;
+        int lowestPitch = Note.MAX_PITCH;
         int currentPitch;
 
         for (int i = 0; i < noteArray.length; i++) {
             currentPitch = noteArray[i].getPitch ();
-            if (currentPitch != (int)jm.JMC.REST) {
+            if (currentPitch != jm.JMC.REST) {
                 if (currentPitch > highestPitch) {
                     highestPitch = currentPitch;
                 } else if (currentPitch < lowestPitch) {
@@ -2060,7 +2060,7 @@ public final class PhraseAnalysis {
                 }
             }
         }
-        if (highestPitch != (int)Note.MIN_PITCH && lowestPitch != (int)Note.MAX_PITCH) {
+        if (highestPitch != Note.MIN_PITCH && lowestPitch != Note.MAX_PITCH) {
             return highestPitch - lowestPitch;
         } else {
             throw new NoteListException ("There are no notes in the melody.");
